+++ LLama/LLamaInteractExecutor.cs
@@ -241,7 +241,7 @@ namespace LLama
 
                 if (IsMultiModal && MtmdChunks is not null)
                 {
-                    var nPast = (long)_pastTokensCount;
+                    var nPast = _pastTokensCount;
                     var previousConsumed = _consumedTokensCount;
                     EvaluateMtmdChunks(ref nPast, previousConsumed, nameof(InteractiveExecutor));
                 }
@@ -262,7 +262,7 @@ namespace LLama
             else if (IsMultiModal && MtmdChunks is not null)
             {
                 _is_prompt_run = false;
-                var nPast = (long)_pastTokensCount;
+                var nPast = _pastTokensCount;
                 var previousConsumed = _consumedTokensCount;
                 EvaluateMtmdChunks(ref nPast, previousConsumed, nameof(InteractiveExecutor));
             }
